component return promise but doesnt render
